Enabled warnings for suspicious implicit casts, and fixed all warnings produced by this. This involved switching specific use of float or double functions to use CGFloat and <tgmath.h>.

+++ BFColorPickerPopover/Extensions/NSColor+BFColorPickerPopover.m
@@ -44,7 +44,7 @@
 }
 
 + (NSColor *)randomColor {
-	return [NSColor colorWithCalibratedRed:(float)random()/RAND_MAX green:(float) random()/RAND_MAX blue:(float) random()/RAND_MAX alpha:1.0f];
+	return [NSColor colorWithCalibratedRed:random()/RAND_MAX green:random()/RAND_MAX blue:random()/RAND_MAX alpha:1.0];
 }
 
 @end
